Refactor server to use SOLID principles by extracting routers and schemas
