+++ .gitignore
@@ -1,3 +1,5 @@
 venv
 brainstorming.md
 repomix-output.md
+walkthrough*
+test*
