+++ refactoring.md
@@ -1,0 +1,29 @@
+# Refactoring History
+
+This document tracks the changes made to the codebase to improve quality and apply SOLID principles.
+
+## Step 1: Extract Rate Limiter (SRP)
+- **Problem**: `server/main.py` was directly responsible for instantiating the rate limiter, cluttering the file.
+- **Solution**: Created a new module `server/app/limiter.py` to instantiate the `Limiter`.
+- **Changes**:
+    - Created `server/app/limiter.py`.
+    - Modified `server/main.py` to import `limiter` from the new module instead of creating it.
+- **Status**: Tested and verified.
+
+## Step 2: Extract Data Schemas and Auth Utilities (SRP)
+- **Problem**: `main.py` contained numerous Pydantic request models and password hashing utilities, which violates the Single Responsibility Principle.
+- **Solution**: Moved all request schemas to `server/app/schemas.py` and password utilities to `server/app/auth.py`.
+- **Changes**:
+    - Appended request models (e.g., `RegisterRequest`, `LoginRequest`) to `schemas.py`.
+    - Appended `verify_password` and `get_password_hash` to `auth.py`.
+    - Updated `main.py` to import these from their new locations instead of defining them inline.
+- **Status**: Tested and verified.
+
+## Step 3 & 4: Modularize Endpoints with APIRouters (SRP & OCP)
+- **Problem**: `main.py` was a "god class" handling all API endpoints for different domains (Auth, Waitlist, Services, Users, Chat), making it large (750+ lines) and hard to maintain.
+- **Solution**: Used FastAPI's `APIRouter` to split routes by domain into a new `routers` directory.
+- **Changes**:
+    - Created `server/app/routers/` directory.
+    - Created `auth.py`, `waitlist.py`, `services.py`, `users.py`, and `chat.py` with their respective endpoints.
+    - Simplified `main.py` to only configure the FastAPI application and include the new routers.
+- **Status**: Tested and verified. The codebase is now highly modular and adheres to SOLID principles.
