+++ server/main.py
@@ -1,41 +1,22 @@
+import os
+from contextlib import asynccontextmanager
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+from slowapi import _rate_limit_exceeded_handler
+from slowapi.errors import RateLimitExceeded
 from dotenv import load_dotenv
+
 load_dotenv() # Load variables from .env file into os.environ
 
-from contextlib import asynccontextmanager
-from fastapi import FastAPI, Depends, HTTPException, status, Request, Response
-from fastapi.middleware.cors import CORSMiddleware
-from sqlmodel import Session, select
-from pydantic import BaseModel, EmailStr, field_validator, Field
-import re
-import os
-import hmac
-import logging
-from slowapi import Limiter, _rate_limit_exceeded_handler
-from slowapi.util import get_remote_address
-from slowapi.errors import RateLimitExceeded
-import bcrypt
-import httpx
-import uuid
-import secrets
-from datetime import date, time, datetime, timedelta
+from app.database import create_db_and_tables
+from app.limiter import limiter
 
-from app.database import create_db_and_tables, get_session
-from app.models import Client, Professional, Waitlist, ServiceRequest, ServiceRequestRejection, ServiceRescheduleProposal
-from app.auth import create_access_token, get_current_user
-from app.schemas import ClientResponse, ProfessionalResponse, ServiceRequestResponse, ServiceRequestPublicResponse, RescheduleProposalResponse
-from app.utils.chatbot import generate_chat_response
-
-
-
-DUMMY_PASSWORD_HASH = os.getenv("DUMMY_PASSWORD_HASH", "$2b$12$EixZaYVK1fsbw1ZfbX3OXePaWxn96p36WQoeG6Lruj3vjIQqiRQYq")
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return bcrypt.checkpw(plain_password.encode('utf-8'), hashed_password.encode('utf-8'))
-
-def get_password_hash(password: str) -> str:
-    return bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt()).decode('utf-8')
-
-limiter = Limiter(key_func=get_remote_address)
+# Import Routers
+from app.routers.auth import router as auth_router
+from app.routers.waitlist import router as waitlist_router
+from app.routers.services import router as services_router
+from app.routers.users import router as users_router
+from app.routers.chat import router as chat_router
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
@@ -55,711 +36,9 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# Pydantic Schemas for Requests
-class RegisterRequest(BaseModel):
-    role: str = Field(max_length=50)
-    name: str = Field(max_length=100)
-    last_name: str = Field(max_length=100)
-    email: EmailStr
-    whatsapp_number: str = Field(max_length=20)
-    password: str = Field(max_length=100)
-    specialty: str | None = Field(default=None, max_length=100)
-    waitlist_id: str
-    first_access_password: str
-
-    @field_validator('password')
-    def validate_password(cls, v):
-        if len(v) < 8:
-            raise ValueError("A senha deve ter no mínimo 8 caracteres.")
-        if not re.search(r"[A-Z]", v):
-            raise ValueError("A senha deve conter pelo menos uma letra maiúscula.")
-        if not re.search(r"[a-z]", v):
-            raise ValueError("A senha deve conter pelo menos uma letra minúscula.")
-        if not re.search(r"\d", v):
-            raise ValueError("A senha deve conter pelo menos um número.")
-        return v
-
-class LoginRequest(BaseModel):
-    email: str
-    password: str
-
-class WaitlistRequest(BaseModel):
-    email: EmailStr
-    phone: str | None = None
-    intended_role: str | None = None
-    requested_service: str | None = None
-
-class AdminLoginRequest(BaseModel):
-    password: str
-
-class ChatMessageItem(BaseModel):
-    role: str
-    content: str
-
-class ChatMessage(BaseModel):
-    history: list[ChatMessageItem]
-
-class FirstAccessRequest(BaseModel):
-    email: EmailStr
-    first_access_password: str
-
-class VerifyEmailRequest(BaseModel):
-    token: str
-
-class ServiceRequestCreate(BaseModel):
-    client_id: uuid.UUID | None = None
-    professional_id: uuid.UUID | None = None
-    service_type: str = Field(max_length=100)
-    home_type: str = Field(max_length=100)
-    bedrooms: str = Field(max_length=50)
-    bathrooms: str = Field(max_length=50)
-    has_pets: bool
-    cep: str = Field(max_length=20)
-    address: str = Field(max_length=255)
-    scheduled_date: date
-    scheduled_time: time
-
-class ServiceRequestUpdate(BaseModel):
-    status: str | None = None
-    professional_id: uuid.UUID | None = None
-    service_type: str | None = None
-    home_type: str | None = None
-    bedrooms: str | None = None
-    bathrooms: str | None = None
-    has_pets: bool | None = None
-    cep: str | None = None
-    address: str | None = None
-    scheduled_date: date | None = None
-    scheduled_time: time | None = None
-
-@app.post("/api/auth/register")
-@limiter.limit("5/minute")
-async def register(request: Request, data: RegisterRequest, session: Session = Depends(get_session)):
-    full_name = f"{data.name} {data.last_name}".strip()
-    hashed_pwd = get_password_hash(data.password)
-    email_lower = data.email.lower()
-    email_token = secrets.token_urlsafe(32)
-    
-    # Secure First Access validation
-    try:
-        w_id = uuid.UUID(data.waitlist_id)
-    except ValueError:
-        raise HTTPException(status_code=400, detail="ID da lista de espera inválido")
-        
-    waitlist_entry = session.get(Waitlist, w_id)
-    if not waitlist_entry or waitlist_entry.first_access_password != data.first_access_password or waitlist_entry.email != email_lower:
-        raise HTTPException(status_code=403, detail="Código de confirmação inválido ou não confere.")
-        
-    if waitlist_entry.is_registered:
-        raise HTTPException(status_code=400, detail="Este convite já foi utilizado para registro.")
-    
-    if data.role == "provider":
-        # Check if email exists
-        existing = session.exec(select(Professional).where(Professional.email == email_lower)).first()
-        if existing:
-            raise HTTPException(status_code=400, detail="Email already registered")
-            
-        new_user = Professional(
-            name=full_name,
-            email=email_lower,
-            whatsapp_number=data.whatsapp_number,
-            password=hashed_pwd,
-            specialty=data.specialty,
-            verification_token=None,
-            email_verified=True,
-            is_verified=False
-        )
-        session.add(new_user)
-    else:
-        # Check if email exists
-        existing = session.exec(select(Client).where(Client.email == email_lower)).first()
-        if existing:
-            raise HTTPException(status_code=400, detail="Email already registered")
-            
-        new_user = Client(
-            name=full_name,
-            email=email_lower,
-            whatsapp_number=data.whatsapp_number,
-            password=hashed_pwd,
-            verification_token=None, 
-            email_verified=True
-        )
-        session.add(new_user)
-        
-    waitlist_entry.is_registered = True
-    session.add(waitlist_entry)
-        
-    session.commit()
-    session.refresh(new_user)
-    
-    try: 
-        n8n_base_url = os.getenv("N8N_WEBHOOK_URL")
-        if not n8n_base_url:
-            raise ValueError("N8N_WEBHOOK_URL environment variable is not defined.")
-        n8n_base_url = n8n_base_url.rstrip('/')
-            
-        async with httpx.AsyncClient() as client:
-            await client.post(f"{n8n_base_url}/webhook/registro-esponja", json={
-                "email": new_user.email,
-                "name": new_user.name,
-                "role": data.role,
-                "verification_token": new_user.verification_token
-            })
-    except Exception as e:
-        print(f"Erro ao disparar gatilho de e-mail no n8n: {e}")
-
-    return {"message": "User created successfully", "role": data.role, "user_id": str(new_user.id)}
-
-@app.post("/api/auth/login")
-@limiter.limit("5/minute")
-def login(request: Request, data: LoginRequest, response: Response, session: Session = Depends(get_session)):
-    email_lower = data.email.lower()
-    admin_email = os.getenv("ADMIN_EMAIL")
-    admin_password = os.getenv("ADMIN_PASSWORD")
-    
-    # Check admin first
-    if admin_email and admin_password and email_lower == admin_email.lower() and hmac.compare_digest(data.password, admin_password):
-        access_token = create_access_token(data={"sub": "admin", "role": "admin"})
-        response.set_cookie(key="access_token", value=access_token, httponly=True, secure=True, samesite="lax")
-        return {"message": "Admin login successful", "role": "admin", "user_id": "admin", "name": "Administrador"}
-
-    # Check clients
-    client = session.exec(select(Client).where(Client.email == email_lower)).first()
-    if client:
-        if verify_password(data.password, client.password):
-            if not client.email_verified: 
-                raise HTTPException(status_code=403, detail="Por favor, verifique seu e-mail antes de fazer login.")
-            access_token = create_access_token(data={"sub": str(client.id), "role": "customer"})
-            response.set_cookie(key="access_token", value=access_token, httponly=True, secure=True, samesite="lax")
-            return {
-                "message": "Login successful", 
-                "role": "customer", 
-                "user_id": str(client.id),
-                "name": client.name
-            }
-    else:
-        # Dummy check to equalize timing for user enumeration prevention
-        verify_password(data.password, DUMMY_PASSWORD_HASH)
-        
-    # Check professionals
-    professional = session.exec(select(Professional).where(Professional.email == email_lower)).first()
-    if professional:
-        if verify_password(data.password, professional.password):
-            if not professional.email_verified: 
-                raise HTTPException(status_code=403, detail="Por favor, verifique seu e-mail antes de fazer login.")
-            access_token = create_access_token(data={"sub": str(professional.id), "role": "provider"})
-            response.set_cookie(key="access_token", value=access_token, httponly=True, secure=True, samesite="lax")
-            return {
-                "message": "Login successful", 
-                "role": "provider", 
-                "user_id": str(professional.id),
-                "name": professional.name
-            }
-    else:
-        # Dummy check to equalize timing for user enumeration prevention
-        verify_password(data.password, DUMMY_PASSWORD_HASH)
-        
-    raise HTTPException(status_code=401, detail="Invalid credentials")
-
-
-@app.post("/api/auth/first-access")
-@limiter.limit("10/minute")
-def first_access(request: Request, data: FirstAccessRequest, session: Session = Depends(get_session)):
-    email_lower = data.email.lower()
-    waitlist_entry = session.exec(select(Waitlist).where(Waitlist.email == email_lower)).first()
-    
-    if not waitlist_entry:
-        raise HTTPException(status_code=404, detail="Email não encontrado na lista de espera.")
-        
-    if waitlist_entry.first_access_password != data.first_access_password:
-        raise HTTPException(status_code=401, detail="Código de confirmação inválido.")
-        
-    if waitlist_entry.is_registered:
-        raise HTTPException(status_code=400, detail="Este email já completou o registro.")
-        
-    return {
-        "message": "Acesso verificado",
-        "waitlist_id": str(waitlist_entry.id),
-        "email": waitlist_entry.email,
-        "phone": waitlist_entry.phone,
-        "intended_role": waitlist_entry.intended_role
-    }
-
-
-@app.post("/api/auth/verify")
-@limiter.limit("5/minute")
-def verify_email(request: Request, data: VerifyEmailRequest, session: Session = Depends(get_session)):
-    client = session.exec(select(Client).where(Client.verification_token == data.token)).first()
-    if client:
-        client.email_verified = True
-        client.verification_token = None
-        session.add(client)
-        session.commit()
-        return {"message": "Email verificado com sucesso!"}
-        
-    professional = session.exec(select(Professional).where(Professional.verification_token == data.token)).first()
-    if professional:
-        professional.email_verified = True
-        professional.verification_token = None 
-        session.add(professional)
-        session.commit()
-        return {"message": "Email verificado com sucesso!"}
-        
-    raise HTTPException(status_code=400, detail="Token inválido ou expirado.")
-
-
-@app.post("/api/auth/logout")
-def logout(response: Response):
-    response.delete_cookie("access_token", httponly=True, secure=True, samesite="lax")
-    return {"message": "Successfully logged out"}
-
-@app.post("/api/waitlist")
-@limiter.limit("10/minute")
-async def join_waitlist(request: Request, data: WaitlistRequest, session: Session = Depends(get_session)):
-    email_lower = data.email.lower()
-    
-    existing_waitlist = session.exec(select(Waitlist).where(Waitlist.email == email_lower)).first()
-    if existing_waitlist:
-        raise HTTPException(status_code=400, detail="Email already on waitlist")
-        
-    existing_client = session.exec(select(Client).where(Client.email == email_lower)).first()
-    if existing_client:
-        raise HTTPException(status_code=400, detail="Email already registered as Client")
-        
-    existing_professional = session.exec(select(Professional).where(Professional.email == email_lower)).first()
-    if existing_professional:
-        raise HTTPException(status_code=400, detail="Email already registered as Professional")
-        
-    first_access_pwd = secrets.token_hex(3)
-    
-    entry = Waitlist(
-        email=email_lower, 
-        phone=data.phone, 
-        intended_role=data.intended_role,
-        requested_service=data.requested_service,
-        first_access_password=first_access_pwd
-    )
-    session.add(entry)
-    session.commit()
-    
-    try: 
-        n8n_base_url = os.getenv("N8N_WEBHOOK_URL")
-        if not n8n_base_url:
-            raise ValueError("N8N_WEBHOOK_URL environment variable is not defined.")
-        n8n_base_url = n8n_base_url.rstrip('/')
-            
-        async with httpx.AsyncClient() as client:
-            response = await client.post(f"{n8n_base_url}/webhook/waitlist-esponja", json={
-                "email": entry.email,
-                "phone": entry.phone,
-                "first_access_password": entry.first_access_password
-            })
-            response.raise_for_status()
-    except httpx.HTTPStatusError as e:
-        logging.error(f"Erro HTTP ao disparar gatilho de e-mail no n8n (waitlist): {e.response.status_code} - {e.response.text}")
-    except httpx.RequestError as e:
-        logging.error(f"Erro de rede ao disparar gatilho de e-mail no n8n (waitlist): {e}")
-    except Exception as e:
-        logging.error(f"Erro inesperado ao disparar gatilho de e-mail no n8n (waitlist): {e}")
-    
-    return {"message": "Successfully added to waitlist"}
-
-# admin_login removed, handled in login endpoint
-@app.post("/api/service-requests")
-@limiter.limit("5/minute")
-def create_service_request(request: Request, data: ServiceRequestCreate, session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    user_id = current_user["user_id"]
-    role = current_user["role"]
-    
-    if role == "customer":
-        data.client_id = uuid.UUID(user_id)
-    elif role == "admin":
-        if not data.client_id:
-            raise HTTPException(status_code=400, detail="Admin must specify client_id")
-    else:
-        raise HTTPException(status_code=403, detail="Providers cannot create requests")
-        
-    client = session.get(Client, data.client_id)
-    if not client:
-        raise HTTPException(status_code=404, detail="Client not found")
-        
-    db_request = ServiceRequest(**data.model_dump())
-    session.add(db_request)
-    session.commit()
-    session.refresh(db_request)
-    return db_request
-
-def get_pending_reschedule_data(session: Session, service_request_id: uuid.UUID) -> dict | None:
-    proposal = session.exec(select(ServiceRescheduleProposal).where(
-        (ServiceRescheduleProposal.service_request_id == service_request_id) &
-        (ServiceRescheduleProposal.status == "Pendente")
-    )).first()
-    if proposal:
-        return {
-            "id": proposal.id,
-            "service_request_id": proposal.service_request_id,
-            "proposed_date": proposal.proposed_date,
-            "proposed_time": proposal.proposed_time,
-            "requested_by_role": proposal.requested_by_role,
-            "status": proposal.status
-        }
-    return None
-
-@app.get("/api/service-requests")
-def get_service_requests(session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    user_id = current_user["user_id"]
-    role = current_user["role"]
-    
-    if role == "admin":
-        requests = session.exec(select(ServiceRequest)).all()
-        result = []
-        for r in requests:
-            d = r.model_dump()
-            if r.professional_id:
-                p = session.get(Professional, r.professional_id)
-                d["professional_name"] = p.name if p else None
-            d["pending_reschedule"] = get_pending_reschedule_data(session, r.id)
-            result.append(ServiceRequestResponse(**d))
-        return result
-    elif role == "customer":
-        requests = session.exec(select(ServiceRequest).where(ServiceRequest.client_id == uuid.UUID(user_id))).all()
-        result = []
-        for r in requests:
-            d = r.model_dump()
-            if r.professional_id:
-                p = session.get(Professional, r.professional_id)
-                d["professional_name"] = p.name if p else None
-            d["pending_reschedule"] = get_pending_reschedule_data(session, r.id)
-            result.append(ServiceRequestResponse(**d))
-        return result
-    else: # provider
-        requests = session.exec(select(ServiceRequest).where(
-            (ServiceRequest.professional_id == uuid.UUID(user_id)) |
-            ((ServiceRequest.status == "Pendente") & (ServiceRequest.professional_id == None))
-        )).all()
-        
-        # Filter out requests rejected by the current professional
-        rejected_ids = set(session.exec(select(ServiceRequestRejection.service_request_id).where(
-            ServiceRequestRejection.professional_id == uuid.UUID(user_id)
-        )).all())
-        
-        result = []
-        for r in requests:
-            if r.id not in rejected_ids:
-                if r.status == "Pendente" and str(r.professional_id) != user_id:
-                    d = r.model_dump()
-                    d["pending_reschedule"] = get_pending_reschedule_data(session, r.id)
-                    result.append(ServiceRequestPublicResponse(**d))
-                else:
-                    d = r.model_dump()
-                    if r.professional_id:
-                        p = session.get(Professional, r.professional_id)
-                        d["professional_name"] = p.name if p else None
-                    d["pending_reschedule"] = get_pending_reschedule_data(session, r.id)
-                    result.append(ServiceRequestResponse(**d))
-        return result
-
-@app.get("/api/service-requests/{request_id}")
-def get_service_request(request_id: uuid.UUID, session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    db_request = session.get(ServiceRequest, request_id)
-    if not db_request:
-        raise HTTPException(status_code=404, detail="Service request not found")
-        
-    user_id = current_user["user_id"]
-    role = current_user["role"]
-    
-    if role == "customer" and str(db_request.client_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-    
-    d = db_request.model_dump()
-    if db_request.professional_id:
-        p = session.get(Professional, db_request.professional_id)
-        d["professional_name"] = p.name if p else None
-    d["pending_reschedule"] = get_pending_reschedule_data(session, db_request.id)
-    return ServiceRequestResponse(**d)
-
-@app.patch("/api/service-requests/{request_id}")
-def update_service_request(request_id: uuid.UUID, update_data: ServiceRequestUpdate, session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    db_request = session.get(ServiceRequest, request_id)
-    if not db_request:
-        raise HTTPException(status_code=404, detail="Service request not found")
-        
-    user_id = current_user["user_id"]
-    role = current_user["role"]
-    
-    # IDOR Check
-    if role == "customer" and str(db_request.client_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-    elif role == "provider":
-        if db_request.professional_id and str(db_request.professional_id) != user_id:
-            raise HTTPException(status_code=403, detail="Not authorized to modify another professional's request")
-        elif not db_request.professional_id:
-            # A provider can only accept a pending request
-            if db_request.status != "Pendente":
-                raise HTTPException(status_code=400, detail="Este serviço já foi aceito por outro profissional.")
-            if update_data.professional_id != uuid.UUID(user_id) or update_data.status != "Em Andamento":
-                raise HTTPException(status_code=403, detail="Providers can only accept pending requests")
-            
-    update_dict = update_data.model_dump(exclude_unset=True)
-    for key, value in update_dict.items():
-        setattr(db_request, key, value)
-        
-    session.add(db_request)
-    session.commit()
-    session.refresh(db_request)
-    return db_request
-
-@app.delete("/api/service-requests/{request_id}")
-def delete_service_request(request_id: uuid.UUID, session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    db_request = session.get(ServiceRequest, request_id)
-    if not db_request:
-        raise HTTPException(status_code=404, detail="Service request not found")
-        
-    user_id = current_user["user_id"]
-    role = current_user["role"]
-    
-    if role == "customer" and str(db_request.client_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-    elif role == "provider":
-        raise HTTPException(status_code=403, detail="Providers cannot delete requests")
-        
-    session.delete(db_request)
-    session.commit()
-    return {"message": "Service request deleted successfully"}
-
-@app.post("/api/service-requests/{request_id}/reject")
-def reject_service_request(request_id: uuid.UUID, session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    if current_user["role"] != "provider":
-        raise HTTPException(status_code=403, detail="Apenas prestadores podem recusar serviços.")
-        
-    db_request = session.get(ServiceRequest, request_id)
-    if not db_request:
-        raise HTTPException(status_code=404, detail="Service request not found")
-        
-    if db_request.status != "Pendente":
-        raise HTTPException(status_code=400, detail="Apenas serviços pendentes podem ser recusados.")
-        
-    user_id = current_user["user_id"]
-    professional_id = uuid.UUID(user_id)
-    
-    # Check if already rejected
-    existing = session.exec(select(ServiceRequestRejection).where(
-        (ServiceRequestRejection.professional_id == professional_id) & 
-        (ServiceRequestRejection.service_request_id == request_id)
-    )).first()
-    
-    if not existing:
-        rejection = ServiceRequestRejection(
-            professional_id=professional_id,
-            service_request_id=request_id
-        )
-        session.add(rejection)
-        session.commit()
-        
-    return {"message": "Service request rejected successfully"}
-
-@app.get("/api/professionals/me", response_model=ProfessionalResponse)
-def get_professional_me(session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    if current_user["role"] != "provider":
-        raise HTTPException(status_code=403, detail="Not authorized")
-    professional = session.get(Professional, uuid.UUID(current_user["user_id"]))
-    if not professional:
-        raise HTTPException(status_code=404, detail="Professional not found")
-    return professional
-
-@app.get("/api/professionals", response_model=list[ProfessionalResponse])
-def get_professionals(session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    if current_user["role"] not in ["admin", "customer"]:
-        raise HTTPException(status_code=403, detail="Not authorized")
-    professionals = session.exec(select(Professional)).all()
-    return professionals
-
-@app.get("/api/professionals/{professional_id}", response_model=ProfessionalResponse)
-def get_professional_by_id(professional_id: uuid.UUID, session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    if current_user["role"] not in ["admin", "customer"]:
-        raise HTTPException(status_code=403, detail="Not authorized")
-    professional = session.get(Professional, professional_id)
-    if not professional:
-        raise HTTPException(status_code=404, detail="Professional not found")
-    return professional
-
-@app.get("/api/clients", response_model=list[ClientResponse])
-def get_clients(
-    session: Session = Depends(get_session),
-    current_user: dict = Depends(get_current_user)
-):
-    if current_user["role"] != "admin":
-        raise HTTPException(status_code=403, detail="Not authorized")
-    clients = session.exec(select(Client)).all()
-    return clients
-
-@app.post("/api/chat")
-@limiter.limit("20/minute")
-def chat_endpoint(request: Request, data: ChatMessage, session: Session = Depends(get_session), current_user: dict = Depends(get_current_user)):
-    if current_user["role"] != "customer":
-        return {"reply": "Você precisa estar logado como cliente para usar o assistente virtual.", "booked": False}
-    result = generate_chat_response(data.history, current_user["user_id"], db_session=session)
-    return {"reply": result["text"], "booked": result["booked"]}
-
-class RescheduleProposalCreate(BaseModel):
-    proposed_date: date
-    proposed_time: time
-
-@app.post("/api/service-requests/{request_id}/reschedule-proposals", response_model=RescheduleProposalResponse)
-def create_reschedule_proposal(
-    request_id: uuid.UUID,
-    data: RescheduleProposalCreate,
-    session: Session = Depends(get_session),
-    current_user: dict = Depends(get_current_user)
-):
-    user_id = current_user["user_id"]
-    role = current_user["role"]
-    
-    if role not in ["customer", "provider"]:
-        raise HTTPException(status_code=403, detail="Not authorized")
-        
-    db_request = session.get(ServiceRequest, request_id)
-    if not db_request:
-        raise HTTPException(status_code=404, detail="Service request not found")
-        
-    if role == "customer" and str(db_request.client_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-    if role == "provider" and str(db_request.professional_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-        
-    service_datetime = datetime.combine(db_request.scheduled_date, db_request.scheduled_time)
-    if service_datetime - datetime.now() < timedelta(hours=24):
-        raise HTTPException(status_code=400, detail="Proposals must be requested at least 24 hours in advance")
-        
-    existing_pending = session.exec(select(ServiceRescheduleProposal).where(
-        (ServiceRescheduleProposal.service_request_id == request_id) &
-        (ServiceRescheduleProposal.status == "Pendente")
-    )).first()
-    if existing_pending:
-        raise HTTPException(status_code=400, detail="There is already a pending reschedule proposal for this service")
-        
-    proposal = ServiceRescheduleProposal(
-        service_request_id=request_id,
-        proposed_date=data.proposed_date,
-        proposed_time=data.proposed_time,
-        requested_by_role=role,
-        status="Pendente"
-    )
-    session.add(proposal)
-    session.commit()
-    session.refresh(proposal)
-    return proposal
-
-
-@app.post("/api/reschedule-proposals/{proposal_id}/accept")
-def accept_reschedule_proposal(
-    proposal_id: uuid.UUID,
-    session: Session = Depends(get_session),
-    current_user: dict = Depends(get_current_user)
-):
-    user_id = current_user["user_id"]
-    role = current_user["role"]
-    
-    if role not in ["customer", "provider"]:
-        raise HTTPException(status_code=403, detail="Not authorized")
-        
-    proposal = session.get(ServiceRescheduleProposal, proposal_id)
-    if not proposal:
-        raise HTTPException(status_code=404, detail="Reschedule proposal not found")
-        
-    if proposal.status != "Pendente":
-        raise HTTPException(status_code=400, detail="Proposal is already resolved")
-        
-    db_request = session.get(ServiceRequest, proposal.service_request_id)
-    if not db_request:
-        raise HTTPException(status_code=404, detail="Service request not found")
-        
-    if proposal.requested_by_role == role:
-        raise HTTPException(status_code=400, detail="You cannot accept your own reschedule proposal")
-        
-    if role == "customer" and str(db_request.client_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-    if role == "provider" and str(db_request.professional_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-        
-    service_datetime = datetime.combine(db_request.scheduled_date, db_request.scheduled_time)
-    if service_datetime - datetime.now() < timedelta(hours=24):
-        raise HTTPException(status_code=400, detail="Actions must be performed at least 24 hours in advance")
-        
-    db_request.scheduled_date = proposal.proposed_date
-    db_request.scheduled_time = proposal.proposed_time
-    session.add(db_request)
-    
-    proposal.status = "Aceito"
-    session.add(proposal)
-    
-    session.commit()
-    return {"message": "Reschedule proposal accepted and updated successfully"}
-
-
-@app.post("/api/reschedule-proposals/{proposal_id}/reject")
-def reject_reschedule_proposal(
-    proposal_id: uuid.UUID,
-    session: Session = Depends(get_session),
-    current_user: dict = Depends(get_current_user)
-):
-    user_id = current_user["user_id"]
-    role = current_user["role"]
-    
-    if role not in ["customer", "provider"]:
-        raise HTTPException(status_code=403, detail="Not authorized")
-        
-    proposal = session.get(ServiceRescheduleProposal, proposal_id)
-    if not proposal:
-        raise HTTPException(status_code=404, detail="Reschedule proposal not found")
-        
-    if proposal.status != "Pendente":
-        raise HTTPException(status_code=400, detail="Proposal is already resolved")
-        
-    db_request = session.get(ServiceRequest, proposal.service_request_id)
-    if not db_request:
-        raise HTTPException(status_code=404, detail="Service request not found")
-        
-    if proposal.requested_by_role == role:
-        raise HTTPException(status_code=400, detail="You cannot reject your own reschedule proposal")
-        
-    if role == "customer" and str(db_request.client_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-    if role == "provider" and str(db_request.professional_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-        
-    proposal.status = "Recusado"
-    session.add(proposal)
-    session.commit()
-    return {"message": "Reschedule proposal rejected successfully"}
-
-
-@app.post("/api/service-requests/{request_id}/cancel")
-def cancel_service_request(
-    request_id: uuid.UUID,
-    session: Session = Depends(get_session),
-    current_user: dict = Depends(get_current_user)
-):
-    user_id = current_user["user_id"]
-    role = current_user["role"]
-    
-    if role not in ["customer", "provider"]:
-        raise HTTPException(status_code=403, detail="Not authorized")
-        
-    db_request = session.get(ServiceRequest, request_id)
-    if not db_request:
-        raise HTTPException(status_code=404, detail="Service request not found")
-        
-    if role == "customer" and str(db_request.client_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-    if role == "provider" and str(db_request.professional_id) != user_id:
-        raise HTTPException(status_code=403, detail="Not authorized")
-        
-    service_datetime = datetime.combine(db_request.scheduled_date, db_request.scheduled_time)
-    if service_datetime - datetime.now() < timedelta(hours=24):
-        raise HTTPException(status_code=400, detail="Cancellations must be performed at least 24 hours in advance")
-        
-    db_request.status = "Cancelado"
-    session.add(db_request)
-    session.commit()
-    return {"message": "Service request cancelled successfully"}
+# Include Routers
+app.include_router(auth_router)
+app.include_router(waitlist_router)
+app.include_router(services_router)
+app.include_router(users_router)
+app.include_router(chat_router)
